# PERF: Cythonize Groupby.cummin/cummax (#15048)