@jreback 2017-01-04
bottleneck will be *way* slower as it will be called in a python loop. The cython grouperby routines have the indexers already and thus everything is a single cython loop.