@MaximilianR 2017-01-04
> bottleneck will be way slower as it will be called in a python loop

Ah, you can't put the bn call in a `with nogil` context manager?