@jreback 2017-01-04
@MaximilianR no, nothing to do with that.

The point of the groupby_* routines is that they take and indexer (already been factorized), and make the results per-group.

bottleneck is for a full 1-d array. Here we have *many* variable sized (the groups) 1-d arrays.