@jreback 2017-01-04
lgtm. pls add a whatsnew note & look at the build log, linting issue with a cython file