@mroeschke 2017-01-05
Added what's new and a test. I don't think my implementation is entirely correct yet so I am expecting the test I added to fail.

Is there a nice way to debug cython code? I'd like to drop a `pdb.set_trace()` to see what's going on.