@mroeschke 2017-01-07
Made the changes related to your comments @jorisvandenbossche, and added a whatsnew & passing tests @jreback. All green.