@mroeschke 2017-01-08
Added additional tests and addressed the the val != val condition @jreback . All green.