@jreback 2017-01-11
thanks @mroeschke 

The other dtypes are fixed in #15054 (e.g. datetime/timedelta).

If you'd do that followup to remove accum would be great!