@mroeschke 2017-01-11
Apologies for not flagging this last night. I made a naive assumption that travis would catch this and the build would fail:

```
In [6]: df #np.int64 type
Out[6]: 
   A                    B
0  1                    3
1  1                    4
2  1 -9223372036854775808
3  1                    2
4  2                    2
5  2                    3
6  2 -9223372036854775808
7  2                    1

In [7]: df.groupby('A').cummin()
Out[7]: 
     B
0  3.0
1  3.0
2  NaN
3  NaN
4  2.0
5  2.0
6  NaN
7  NaN

```

Thanks for pointing out the root issue. I'll look into addressing it!